crews.go

+++ crews.go
@@ -41,6 +41,6 @@ func (c *Crews) PrimaryTest(s string) bool {
 func (c *Crews) SecondaryTest(s string) bool {
 	count := c.secondary.Test([]byte(s))
 	adjustedCount := uint((float64(count) - c.p * float64(c.secondary.count))/(c.q-c.p))
-	return  adjustedCount >= uint((c.secondary.count/c.numWebsites)*c.secondaryThresholdModifer)
+	return  adjustedCount >= uint(float64((c.secondary.count/c.numWebsites))*c.secondaryThresholdModifer)
 }
